:recycle: moves files around
