+++ src/routes/rdpc-proxy/handlers/entitiesIdHandler.ts
@@ -2,30 +2,29 @@ import { Client } from '@elastic/elasticsearch';
 import { Request, Response, Handler } from 'express';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import logger from 'utils/logger';
-
+import { hasSufficientProgramMembershipAccess } from '../accessValidations';
 
 const normalizePath = (rootPath: string) => (pathName: string, req: Request) =>
   pathName.replace(rootPath, '').replace('//', '/');
 
-const createEntitiesHandler = ({
+const createEntitiesIdHandler = ({
   esClient,
   rootPath,
 }: {
   esClient: Client;
   rootPath: string;
 }): Handler => {
-  return async (
-      req: Request<{}, any, any, { gnosId: string; size: string; page: string }>,
-      res,
-      next,
-    ) => {
-      /**
-       * @todo: actually implement the API at https://song.rdpc-dev.cancercollaboratory.org/swagger-ui.html#/
-       */
-      const {
-        authorization
-      } = req.headers
-      const egoJwtOrApiKey = (authorization || "")?.split('Bearer ').join('')
+  return async (req: Request<{ fileObjectId: string }>, res, next) => {
+    /**
+     * @todo: actually implement the API at https://song.rdpc-dev.cancercollaboratory.org/swagger-ui.html#/
+     */
+    const { authorization } = req.headers;
+    const egoJwtOrApiKey = (authorization || '')?.split('Bearer ').join('');
+    const isAuthorized = hasSufficientProgramMembershipAccess({
+      egoJwtOrApiKey,
+      file: undefined,
+    });
+    if (isAuthorized) {
       // this is just a placeholder logic for demo
       const handleRequest = createProxyMiddleware({
         target: 'https://song.rdpc-dev.cancercollaboratory.org',
@@ -36,8 +35,11 @@ const createEntitiesHandler = ({
         },
         changeOrigin: true,
       });
-      handleRequest(req, res, next);        
+      handleRequest(req, res, next);
+    } else {
+      res.status(403)
     }
-}
+  };
+};
 
-export default createEntitiesHandler
+export default createEntitiesIdHandler;
